Common. Move Settings and IniConfig
get_settings and storage functions
